+++ example/integration_test/src/tests/menu/locale_menu.dart
@@ -30,7 +30,8 @@ Future<void> openLocaleMenu(WidgetTester tester) async {
 
 Future<void> selectLanguage(WidgetTester tester) async {
   //
-  final listFinder = find.byType(Scrollable, skipOffstage: false);
+//  final listFinder = find.byType(Scrollable); //, skipOffstage: false);
+  final listFinder = find.byKey(const Key('LocaleCupertinoSpinner'));
 
   expect(listFinder, findsWidgets, reason: _location);