+++ example/lib/src/view.dart
@@ -1,8 +1,8 @@
 // The fluttery framework.
-export 'package:fluttery_framework/view.dart' hide AppDrawer, AppMenu;
+export 'package:fluttery_framework/view.dart' hide AppDrawer;
 
 // The UI at the app level.
-export 'package:fluttery_framework_example/src/app/view.dart' hide ISOSpinner;
+export 'package:fluttery_framework_example/src/app/view.dart';
 
 // The UI for the home screen.
 export 'package:fluttery_framework_example/src/home/view.dart';